refactor: DefaultRepository Refactor Data Layer 로 이동
